Feat: Feed 엔티티에 Clothes 매핑 추가 및 likeCount 타입 수정
- Feed와 Clothes 간 다대다 매핑 관계 추가
- Feed의 likeCount 필드 타입을 int에서 long으로 변경
- schema.sql에서 like_count 칼럼 타입 bigint로 수정

+++ src/main/resources/schema.sql
@@ -52,7 +52,7 @@ CREATE TABLE feeds (
                          created_at	timestamp		NOT NULL,
                          updated_at	timestamp		NULL,
                          content	varchar(500)		NOT NULL,
-                         like_count	integer	DEFAULT 0	NOT NULL,
+                         like_count	bigint	DEFAULT 0	NOT NULL,
                          comment_count	integer	DEFAULT 0	NOT NULL,
                          author_id	uuid		NULL,
